Implement a comparison of submitted documents to find nearly-exact submissions.
Add the StringSimilarity class with utility methods to compare two
strings and return a similarity ratio between the strings. Currently,
the Levenshtein distance is used to compute the edit distance between
the strings.
Change the AssignmentResults class to collect the text from the "other"
files in a map so we can re-use the text strings for comparison between
student's submissions.

Add findClosestMatchesInAllSubmissions() method to AssignmentResults to
compare all "other" files in this submission to all other submissions
using the StringSimilarity methods. Collect the most similar file(s) in
the new closestOtherSubmittedFiles map.

In the AssignmentResults toString() method, show the most similar files.
Also, reorder output so the submitted documents are shown before the
Java code to ease grading, since this the docs should be the answer to
the first homework question.